+++ game/sprites/vegetation.py
@@ -1,6 +1,7 @@
 from pygame import sprite, image, transform
 import random
 import math
+
 class Vegetation(sprite.Sprite):
     """ Vegetation sprite class for creating and updating the vegetation in the game screen
     """
@@ -15,11 +16,9 @@ class Vegetation(sprite.Sprite):
         pos = 0
         for _ in range(math.ceil(self.surf.get_width() / ground.get_width())):
             self.surf.blit(ground, (pos, self.surf.get_height() - 40))
-            if random.choice([0,1]) == 0:
-                self.surf.blit(grass, (pos-38, self.surf.get_height() -40-32))
+            if random.choice([True,False]):
+                self.surf.blit(grass, (pos-38, self.surf.get_height() -40 -28))
             pos += ground.get_width()
-
-        
         x_pos = self.__game_env.static.screen_width if x_pos is None else x_pos
         y_pos = self.__game_env.static.screen_height - self.surf.get_height() / 2 if y_pos is None else y_pos
         self.rect = self.surf.get_rect(center=(x_pos, y_pos))