+++ pybluesky.py
@@ -37,6 +37,7 @@ from game.sprites.jet import Jet
 from game.sprites.missile import Missile
 from game.sprites.cloud import Cloud
 from game.sprites.vegetation import Vegetation
+from game.sprites.samlauncher import SamLauncher
 from game.sprites.star import Star
 from game.sprites.text import Text
 from game.sprites.text.score import ScoreText
@@ -55,7 +56,7 @@ API_KEY = ''
 def load_leaders():
     leader_thread = Thread(target=LeaderBoardHandler().update, args=([API_KEY]))
     leader_thread.start()
-    leader_thread.join()
+    leader_thread.join()  
 
 def check_update(game_env):
     network_handler = NetworkHandler(API_KEY)
@@ -88,6 +89,7 @@ def play():
     game_env.dynamic.shoot_sound.set_volume(1.5)
     game_env.dynamic.hit_sound.set_volume(3)
     game_env.dynamic.powerup_sound.set_volume(10)
+    game_env.dynamic.samfire_sound.set_volume(5)
 
     pygame.mixer.music.load(game_env.static.game_sound.get('music'))                                        # setting main game background musicm
     pygame.mixer.music.play(loops=-1)                                                                       # lopping the main game music
@@ -108,19 +110,23 @@ def play():
     ADD_CLOUD = pygame.USEREVENT + 2                                                    # creating custom event to automatically add cloud in the screen
     pygame.time.set_timer(ADD_CLOUD, int(1000/game_env.static.cloud_per_sec))           # setting event to auto-trigger every 1s; 1 cloud will be created every second
 
+    ADD_SAM_LAUNCHER = pygame.USEREVENT + 3
+    pygame.time.set_timer(ADD_SAM_LAUNCHER, 5000)                                       # setting event to auto-trigger every 5s; 1 level can have 4 sam launcher
+
     running = True                                                                      # game running variable
     gameover = False                                                                    # no gameover by default
     game_started = False                                                                # game is not started by default
     star_shown = False
-    mouse_pos = (game_env.static.screen_width, game_env.static.screen_height/2)   # default mouse position, let the jet move forward on a straight line
+    mouse_pos = (game_env.static.screen_width, game_env.static.screen_height/2)         # default mouse position, let the jet move forward on a straight line
     screen_color = game_env.static.background_default if game_started else game_env.static.background_special
 
     backgrounds = pygame.sprite.Group()                                                 # creating seperate group for background sprites
     stars = pygame.sprite.GroupSingle()                                                 # group of stars with max 1 sprite
     vegetations = pygame.sprite.Group()                                                 # creating cloud group for storing all the clouds in the game
     clouds = pygame.sprite.Group()                                                      # creating cloud group for storing all the clouds in the game
-    missiles = pygame.sprite.Group()                                                    # creating missile group for storing all teh missiles in the game
-    deactivated_missile = pygame.sprite.Group()                                         # creating missile group for storing all teh missiles in the game
+    missiles = pygame.sprite.Group()                                                    # creating missile group for storing all the missiles in the game
+    deactivated_missile = pygame.sprite.Group()                                         # creating missile group for storing all the deactivated missiles in the game
+    samlaunchers = pygame.sprite.GroupSingle()                                          # creating missile group for storing all the samlaunchers in the game
     title_sprites = pygame.sprite.Group()
 
     general_hint_text = "Press M->Game menu, H->Help, L->LeaderBoard, ESC->Quit"
@@ -157,6 +163,10 @@ def play():
                     new_missile = Missile(game_env)                                                                 # create a new missile
                     missiles.add(new_missile)                                                                       # adding the missile to missle group
                     game_env.dynamic.all_sprites.add(new_missile)                                                   # adding the missile to all_sprites group as well
+                if event.type == ADD_SAM_LAUNCHER and not samlaunchers.sprites() and game_env.dynamic.game_level > 5:
+                    samlauncher = SamLauncher(game_env)
+                    samlaunchers.add(samlauncher)
+                    game_env.dynamic.all_sprites.add(samlauncher)
             elif event.type == game_env.KEYDOWN:                                                                    # handling all the VALID key press, action varies based on current active screen
                 if not game_started and game_env.dynamic.active_screen == Screen.NAMEINPUT:
                     active_sprite.render(event.unicode)
@@ -178,10 +188,10 @@ def play():
                     game_env.dynamic.active_screen = Screen.LEADERBOARD
                 elif event.key == game_env.K_m:
                     game_env.dynamic.all_sprites.remove(active_sprite)
-                    if not gameover and game_env.dynamic.active_screen != Screen.GAMEMENU:                     # displyaing the game menu
+                    if not gameover and game_env.dynamic.active_screen != Screen.GAMEMENU:                          # displyaing the game menu
                         active_sprite = GameMenuText(game_env)
                         game_env.dynamic.active_screen = Screen.GAMEMENU
-                    elif gameover and game_env.dynamic.active_screen != Screen.REPLAYMENU:                     # displaying the replay menu
+                    elif gameover and game_env.dynamic.active_screen != Screen.REPLAYMENU:                          # displaying the replay menu
                         active_sprite = ReplayText(game_env)
                         game_env.dynamic.active_screen = Screen.REPLAYMENU
                     game_env.dynamic.all_sprites.add(active_sprite)
@@ -234,26 +244,26 @@ def play():
                         pygame.time.set_timer(ADD_MISSILE, int(1000/(game_env.static.missile_per_sec + int(game_env.dynamic.game_level/2)))) # updating timer of ADD_MISSLE for more missiles to be added
                         game_env.dynamic.ammo += 50                                                 # adding 50 ammo on each level up
                         game_env.dynamic.game_score += 10                                           # increasing game score by 10 after each level
-                        game_env.dynamic.all_sprites.remove(game_env.dynamic.noammo_sprite)         # removing no ammo sprite when ammo is refilled
-                        
+                        game_env.dynamic.all_sprites.remove(game_env.dynamic.noammo_sprite)         # removing no ammo sprite when ammo is refilled                       
 
         screen.fill(screen_color)                                                                   # Filling screen with sky blue color
         [screen.blit(sprite.surf, sprite.rect) for sprite in backgrounds]                           # drawing all backgrounds sprites
         [screen.blit(sprite.surf, sprite.rect) for sprite in game_env.dynamic.all_sprites]          # drawing all sprites in the screen
 
         # missile hit
-        if pygame.sprite.spritecollideany(jet, missiles) and not gameover:                          # Check if any missiles have collided with the player; if so
+        if pygame.sprite.spritecollideany(jet, missiles) or pygame.sprite.spritecollideany(jet, game_env.dynamic.sam_missiles) and not gameover:    # Check if any missiles have collided with the player; if so
             active_sprite = ReplayText(game_env)
             game_env.dynamic.active_screen = Screen.REPLAYMENU
             jet.kill()                                                                              # killing the jet
+            [sam_missile.kill() for sam_missile in game_env.dynamic.sam_missiles]                   # killing the SAM missile
             game_env.dynamic.collision_sound.play()
             [game_env.dynamic.all_sprites.add(sprite) for sprite in (active_sprite, hint_sprite)]   # adding the gameover and the hint sprite
             game_env.dynamic.all_sprites.remove(game_env.dynamic.noammo_sprite)
             gameover = True                                                                         # setting gameover to true to prevent new missiles from spawning
             submit_result(game_env)
-        
+
         # missile hit
-        collision = pygame.sprite.groupcollide(missiles, game_env.dynamic.bullets, True, True)      # checking for collision between bullets and missiles, killing each one of them on collision
+        collision = pygame.sprite.groupcollide(missiles, game_env.dynamic.bullets, True, True) # checking for collision between bullets and missiles, killing each one of them on collision
         if len(collision) > 0 and not gameover:
             game_env.dynamic.hit_sound.play()                                                       # play missile destroyed sound
             game_env.dynamic.game_score += len(collision) * 10                                      # 1 missle destroyed = 10 pts.
@@ -288,10 +298,12 @@ def play():
             vegetations.update()                                                                    # vegetations will move only after the game starts
 
         game_env.dynamic.bullets.update()
+        game_env.dynamic.sam_missiles.update()
         missiles.update()                                                                           # update the position of the missiles
         deactivated_missile.update()
         clouds.update()                                                                             # update the postition of the clouds
         stars.update()
+        samlaunchers.update((jet.rect.x + jet.rect.width/2, jet.rect.y+jet.rect.height))
         scoretext_sprite.update()                                                                   # update the game score
 
     pygame.mixer.music.stop()                                                                       # stopping game music